move temp to live
